Refactored counter implementation
